+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking Changes
 - program(chainlink-datastreams): Record last update diff in seconds instead of nanoseconds.
 
+### Fixed
+
+- sdk(sdk): Ensured pay token account is prepared for wrap native.
+
 ### Changes
 
 - programs(utils): Supported `last_update_diff` in seconds in `PriceFeedPrice`.