+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.1] - 2025-08-15
+
 ### Breaking Changes
 - program(chainlink-datastreams): Record last update diff in seconds instead of nanoseconds.
 
@@ -321,7 +323,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Provided SDK (`gmsol`) and other utility crates.
 - Provided a command-line interface (`gmsol`).
 
-[unreleased]: https://github.com/gmsol-labs/gmx-solana/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/gmsol-labs/gmx-solana/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.7.1
 [0.7.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.7.0
 [0.6.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.6.0
 [0.5.0]: https://github.com/gmsol-labs/gmx-solana/releases/tag/v0.5.0